+++ runtime/memory_usage.cpp
@@ -1,13 +1,13 @@
 #include "runtime/memory_usage.h"
 
-int f$estimate_memory_usage(const string &value) {
+int64_t f$estimate_memory_usage(const string &value) {
   if (value.is_reference_counter(ExtraRefCnt::for_global_const) || value.is_reference_counter(ExtraRefCnt::for_instance_cache)) {
     return 0;
   }
-  return static_cast<int>(value.estimate_memory_usage());
+  return static_cast<int64_t>(value.estimate_memory_usage());
 }
 
-int f$estimate_memory_usage(const var &value) {
+int64_t f$estimate_memory_usage(const var &value) {
   if(value.is_string()) {
     return f$estimate_memory_usage(value.as_string());
   } else if (value.is_array()) {
